+++ src/StackExchange.Redis/RedisDatabase.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Buffers;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Net;
 using System.Threading.Tasks;
 using Pipelines.Sockets.Unofficial.Arenas;
@@ -2423,7 +2424,34 @@ namespace StackExchange.Redis
             return ExecuteAsync(msg, ResultProcessor.Int64);
         }
 
-        public RedisValue StreamAdd(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId = null, int? maxLength = null, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None)
+        public StreamTrimResult StreamAcknowledgeAndDelete(RedisKey key, RedisValue groupName, StreamTrimMode mode, RedisValue messageId, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamAcknowledgeAndDeleteMessage(key, groupName, mode, messageId, flags);
+            return ExecuteSync(msg, ResultProcessor.StreamTrimResult);
+        }
+
+        public Task<StreamTrimResult> StreamAcknowledgeAndDeleteAsync(RedisKey key, RedisValue groupName, StreamTrimMode mode, RedisValue messageId, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamAcknowledgeAndDeleteMessage(key, groupName, mode, messageId, flags);
+            return ExecuteAsync(msg, ResultProcessor.StreamTrimResult);
+        }
+
+        public StreamTrimResult[] StreamAcknowledgeAndDelete(RedisKey key, RedisValue groupName, StreamTrimMode mode, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamAcknowledgeAndDeleteMessage(key, groupName, mode, messageIds, flags);
+            return ExecuteSync(msg, ResultProcessor.StreamTrimResultArray)!;
+        }
+
+        public Task<StreamTrimResult[]> StreamAcknowledgeAndDeleteAsync(RedisKey key, RedisValue groupName, StreamTrimMode mode, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamAcknowledgeAndDeleteMessage(key, groupName, mode, messageIds, flags);
+            return ExecuteAsync(msg, ResultProcessor.StreamTrimResultArray)!;
+        }
+
+        public RedisValue StreamAdd(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId, int? maxLength, bool useApproximateMaxLength, CommandFlags flags)
+            => StreamAdd(key, streamField, streamValue, messageId, maxLength, useApproximateMaxLength, null, StreamTrimMode.KeepReferences, flags);
+
+        public RedisValue StreamAdd(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId = null, long? maxLength = null, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
         {
             var msg = GetStreamAddMessage(
                 key,
@@ -2431,12 +2459,17 @@ namespace StackExchange.Redis
                 maxLength,
                 useApproximateMaxLength,
                 new NameValueEntry(streamField, streamValue),
+                limit,
+                mode,
                 flags);
 
             return ExecuteSync(msg, ResultProcessor.RedisValue);
         }
 
-        public Task<RedisValue> StreamAddAsync(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId = null, int? maxLength = null, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None)
+        public Task<RedisValue> StreamAddAsync(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId, int? maxLength, bool useApproximateMaxLength, CommandFlags flags)
+                => StreamAddAsync(key, streamField, streamValue, messageId, maxLength, useApproximateMaxLength, null, StreamTrimMode.KeepReferences, flags);
+
+        public Task<RedisValue> StreamAddAsync(RedisKey key, RedisValue streamField, RedisValue streamValue, RedisValue? messageId = null, long? maxLength = null, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
         {
             var msg = GetStreamAddMessage(
                 key,
@@ -2444,12 +2477,17 @@ namespace StackExchange.Redis
                 maxLength,
                 useApproximateMaxLength,
                 new NameValueEntry(streamField, streamValue),
+                limit,
+                mode,
                 flags);
 
             return ExecuteAsync(msg, ResultProcessor.RedisValue);
         }
 
-        public RedisValue StreamAdd(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId = null, int? maxLength = null, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None)
+        public RedisValue StreamAdd(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId, int? maxLength, bool useApproximateMaxLength, CommandFlags flags)
+            => StreamAdd(key, streamPairs, messageId, maxLength, useApproximateMaxLength, null, StreamTrimMode.KeepReferences, flags);
+
+        public RedisValue StreamAdd(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId = null, long? maxLength = null, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
         {
             var msg = GetStreamAddMessage(
                 key,
@@ -2457,12 +2495,17 @@ namespace StackExchange.Redis
                 maxLength,
                 useApproximateMaxLength,
                 streamPairs,
+                limit,
+                mode,
                 flags);
 
             return ExecuteSync(msg, ResultProcessor.RedisValue);
         }
 
-        public Task<RedisValue> StreamAddAsync(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId = null, int? maxLength = null, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None)
+        public Task<RedisValue> StreamAddAsync(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId, int? maxLength, bool useApproximateMaxLength, CommandFlags flags)
+            => StreamAddAsync(key, streamPairs, messageId, maxLength, useApproximateMaxLength, null, StreamTrimMode.KeepReferences, flags);
+
+        public Task<RedisValue> StreamAddAsync(RedisKey key, NameValueEntry[] streamPairs, RedisValue? messageId = null, long? maxLength = null, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
         {
             var msg = GetStreamAddMessage(
                 key,
@@ -2470,6 +2513,8 @@ namespace StackExchange.Redis
                 maxLength,
                 useApproximateMaxLength,
                 streamPairs,
+                limit,
+                mode,
                 flags);
 
             return ExecuteAsync(msg, ResultProcessor.RedisValue);
@@ -2703,26 +2748,48 @@ namespace StackExchange.Redis
 
         public long StreamDelete(RedisKey key, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None)
         {
-            var msg = Message.Create(
-                Database,
-                flags,
-                RedisCommand.XDEL,
-                key,
-                messageIds);
-
+            var msg = Message.Create(Database, flags, RedisCommand.XDEL, key, messageIds);
             return ExecuteSync(msg, ResultProcessor.Int64);
+        }
+
+        public StreamTrimResult[] StreamDelete(RedisKey key, RedisValue[] messageIds, StreamTrimMode mode, CommandFlags flags)
+        {
+            var msg = GetStreamDeleteExMessage(key, messageIds, mode, flags);
+            return ExecuteSync(msg, ResultProcessor.StreamTrimResultArray)!;
+        }
+
+        private Message GetStreamDeleteExMessage(RedisKey key, RedisValue[] messageIds, StreamTrimMode mode, CommandFlags flags)
+        {
+            if (messageIds == null) throw new ArgumentNullException(nameof(messageIds));
+            if (messageIds.Length == 0) throw new ArgumentOutOfRangeException(nameof(messageIds), "messageIds must contain at least one item.");
+
+            // avoid array for single message case
+            if (messageIds.Length == 1)
+            {
+                return Message.Create(Database, flags, RedisCommand.XDELEX, key, StreamConstants.GetMode(mode), StreamConstants.Ids, 1, messageIds[0]);
+            }
+
+            var values = new RedisValue[messageIds.Length + 3];
+
+            var offset = 0;
+            values[offset++] = StreamConstants.GetMode(mode);
+            values[offset++] = StreamConstants.Ids;
+            values[offset++] = messageIds.Length;
+            messageIds.AsSpan().CopyTo(values.AsSpan(offset));
+            Debug.Assert(offset + messageIds.Length == values.Length);
+            return Message.Create(Database, flags, RedisCommand.XDELEX, key, values);
         }
 
         public Task<long> StreamDeleteAsync(RedisKey key, RedisValue[] messageIds, CommandFlags flags = CommandFlags.None)
         {
-            var msg = Message.Create(
-                Database,
-                flags,
-                RedisCommand.XDEL,
-                key,
-                messageIds);
-
+            var msg = Message.Create(Database, flags, RedisCommand.XDEL, key, messageIds);
             return ExecuteAsync(msg, ResultProcessor.Int64);
+        }
+
+        public Task<StreamTrimResult[]> StreamDeleteAsync(RedisKey key, RedisValue[] messageIds, StreamTrimMode mode, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamDeleteExMessage(key, messageIds, mode, flags);
+            return ExecuteAsync(msg, ResultProcessor.StreamTrimResultArray)!;
         }
 
         public long StreamDeleteConsumer(RedisKey key, RedisValue groupName, RedisValue consumerName, CommandFlags flags = CommandFlags.None)
@@ -2995,15 +3062,33 @@ namespace StackExchange.Redis
             return ExecuteAsync(msg, ResultProcessor.MultiStream, defaultValue: Array.Empty<RedisStream>());
         }
 
-        public long StreamTrim(RedisKey key, int maxLength, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None)
+        public long StreamTrim(RedisKey key, int maxLength, bool useApproximateMaxLength, CommandFlags flags)
+            => StreamTrim(key, maxLength, useApproximateMaxLength, null, StreamTrimMode.KeepReferences, flags);
+
+        public long StreamTrim(RedisKey key, long maxLength, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
         {
-            var msg = GetStreamTrimMessage(key, maxLength, useApproximateMaxLength, flags);
+            var msg = GetStreamTrimMessage(true, key, maxLength, useApproximateMaxLength, limit, mode, flags);
             return ExecuteSync(msg, ResultProcessor.Int64);
         }
 
-        public Task<long> StreamTrimAsync(RedisKey key, int maxLength, bool useApproximateMaxLength = false, CommandFlags flags = CommandFlags.None)
+        public Task<long> StreamTrimAsync(RedisKey key, int maxLength, bool useApproximateMaxLength, CommandFlags flags)
+            => StreamTrimAsync(key, maxLength, useApproximateMaxLength, null, StreamTrimMode.KeepReferences, flags);
+
+        public Task<long> StreamTrimAsync(RedisKey key, long maxLength, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
         {
-            var msg = GetStreamTrimMessage(key, maxLength, useApproximateMaxLength, flags);
+            var msg = GetStreamTrimMessage(true, key, maxLength, useApproximateMaxLength, limit, mode, flags);
+            return ExecuteAsync(msg, ResultProcessor.Int64);
+        }
+
+        public long StreamTrimByMinId(RedisKey key, RedisValue minId, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamTrimMessage(false, key, minId, useApproximateMaxLength, limit, mode, flags);
+            return ExecuteSync(msg, ResultProcessor.Int64);
+        }
+
+        public Task<long> StreamTrimByMinIdAsync(RedisKey key, RedisValue minId, bool useApproximateMaxLength = false, long? limit = null, StreamTrimMode mode = StreamTrimMode.KeepReferences, CommandFlags flags = CommandFlags.None)
+        {
+            var msg = GetStreamTrimMessage(false, key, minId, useApproximateMaxLength, limit, mode, flags);
             return ExecuteAsync(msg, ResultProcessor.Int64);
         }
 
@@ -4109,13 +4194,7 @@ namespace StackExchange.Redis
 
         private Message GetStreamAcknowledgeMessage(RedisKey key, RedisValue groupName, RedisValue messageId, CommandFlags flags)
         {
-            var values = new RedisValue[]
-            {
-                groupName,
-                messageId,
-            };
-
-            return Message.Create(Database, flags, RedisCommand.XACK, key, values);
+            return Message.Create(Database, flags, RedisCommand.XACK, key, groupName, messageId);
         }
 
         private Message GetStreamAcknowledgeMessage(RedisKey key, RedisValue groupName, RedisValue[] messageIds, CommandFlags flags)
@@ -4124,27 +4203,45 @@ namespace StackExchange.Redis
             if (messageIds.Length == 0) throw new ArgumentOutOfRangeException(nameof(messageIds), "messageIds must contain at least one item.");
 
             var values = new RedisValue[messageIds.Length + 1];
-
-            var offset = 0;
-
-            values[offset++] = groupName;
-
-            for (var i = 0; i < messageIds.Length; i++)
-            {
-                values[offset++] = messageIds[i];
-            }
+            values[0] = groupName;
+            messageIds.AsSpan().CopyTo(values.AsSpan(1));
 
             return Message.Create(Database, flags, RedisCommand.XACK, key, values);
         }
 
-        private Message GetStreamAddMessage(RedisKey key, RedisValue messageId, int? maxLength, bool useApproximateMaxLength, NameValueEntry streamPair, CommandFlags flags)
+        private Message GetStreamAcknowledgeAndDeleteMessage(RedisKey key, RedisValue groupName, StreamTrimMode mode, RedisValue messageId, CommandFlags flags)
+        {
+            return Message.Create(Database, flags, RedisCommand.XACKDEL, key, groupName, StreamConstants.GetMode(mode), StreamConstants.Ids, 1, messageId);
+        }
+
+        private Message GetStreamAcknowledgeAndDeleteMessage(RedisKey key, RedisValue groupName, StreamTrimMode mode, RedisValue[] messageIds, CommandFlags flags)
+        {
+            if (messageIds == null) throw new ArgumentNullException(nameof(messageIds));
+            if (messageIds.Length == 0) throw new ArgumentOutOfRangeException(nameof(messageIds), "messageIds must contain at least one item.");
+
+            var values = new RedisValue[messageIds.Length + 4];
+
+            var offset = 0;
+            values[offset++] = groupName;
+            values[offset++] = StreamConstants.GetMode(mode);
+            values[offset++] = StreamConstants.Ids;
+            values[offset++] = messageIds.Length;
+            messageIds.AsSpan().CopyTo(values.AsSpan(offset));
+            Debug.Assert(offset + messageIds.Length == values.Length);
+
+            return Message.Create(Database, flags, RedisCommand.XACKDEL, key, values);
+        }
+
+        private Message GetStreamAddMessage(RedisKey key, RedisValue messageId, long? maxLength, bool useApproximateMaxLength, NameValueEntry streamPair, long? limit, StreamTrimMode mode, CommandFlags flags)
         {
             // Calculate the correct number of arguments:
             //  3 array elements for Entry ID & NameValueEntry.Name & NameValueEntry.Value.
             //  2 elements if using MAXLEN (keyword & value), otherwise 0.
             //  1 element if using Approximate Length (~), otherwise 0.
             var totalLength = 3 + (maxLength.HasValue ? 2 : 0)
-                                + (maxLength.HasValue && useApproximateMaxLength ? 1 : 0);
+                                + (maxLength.HasValue && useApproximateMaxLength ? 1 : 0)
+                                + (limit.HasValue ? 2 : 0)
+                                + (mode != StreamTrimMode.KeepReferences ? 1 : 0);
 
             var values = new RedisValue[totalLength];
             var offset = 0;
@@ -4156,26 +4253,35 @@ namespace StackExchange.Redis
                 if (useApproximateMaxLength)
                 {
                     values[offset++] = StreamConstants.ApproximateMaxLen;
-                    values[offset++] = maxLength.Value;
                 }
-                else
-                {
-                    values[offset++] = maxLength.Value;
-                }
+
+                values[offset++] = maxLength.Value;
+            }
+
+            if (limit.HasValue)
+            {
+                values[offset++] = RedisLiterals.LIMIT;
+                values[offset++] = limit.Value;
+            }
+
+            if (mode != StreamTrimMode.KeepReferences)
+            {
+                values[offset++] = StreamConstants.GetMode(mode);
             }
 
             values[offset++] = messageId;
 
             values[offset++] = streamPair.Name;
-            values[offset] = streamPair.Value;
+            values[offset++] = streamPair.Value;
 
+            Debug.Assert(offset == totalLength);
             return Message.Create(Database, flags, RedisCommand.XADD, key, values);
         }
 
         /// <summary>
         /// Gets message for <see href="https://redis.io/commands/xadd"/>.
         /// </summary>
-        private Message GetStreamAddMessage(RedisKey key, RedisValue entryId, int? maxLength, bool useApproximateMaxLength, NameValueEntry[] streamPairs, CommandFlags flags)
+        private Message GetStreamAddMessage(RedisKey key, RedisValue entryId, long? maxLength, bool useApproximateMaxLength, NameValueEntry[] streamPairs, long? limit, StreamTrimMode mode, CommandFlags flags)
         {
             if (streamPairs == null) throw new ArgumentNullException(nameof(streamPairs));
             if (streamPairs.Length == 0) throw new ArgumentOutOfRangeException(nameof(streamPairs), "streamPairs must contain at least one item.");
@@ -4209,6 +4315,17 @@ namespace StackExchange.Redis
                 values[offset++] = maxLength.Value;
             }
 
+            if (limit.HasValue)
+            {
+                values[offset++] = RedisLiterals.LIMIT;
+                values[offset++] = limit.Value;
+            }
+
+            if (mode != StreamTrimMode.KeepReferences)
+            {
+                values[offset++] = StreamConstants.GetMode(mode);
+            }
+
             values[offset++] = entryId;
 
             for (var i = 0; i < streamPairs.Length; i++)
@@ -4217,6 +4334,7 @@ namespace StackExchange.Redis
                 values[offset++] = streamPairs[i].Value;
             }
 
+            Debug.Assert(offset == totalLength);
             return Message.Create(Database, flags, RedisCommand.XADD, key, values);
         }
 
@@ -4465,26 +4583,44 @@ namespace StackExchange.Redis
             public override int ArgCount => argCount;
         }
 
-        private Message GetStreamTrimMessage(RedisKey key, int maxLength, bool useApproximateMaxLength, CommandFlags flags)
+        private Message GetStreamTrimMessage(bool maxLen, RedisKey key, RedisValue threshold, bool useApproximateMaxLength, long? limit, StreamTrimMode mode, CommandFlags flags)
         {
-            if (maxLength < 0)
+            if (limit.HasValue && limit.GetValueOrDefault() <= 0)
             {
-                throw new ArgumentOutOfRangeException(nameof(maxLength), "maxLength must be equal to or greater than 0.");
+                throw new ArgumentOutOfRangeException(nameof(limit), "limit must be greater than 0 when specified.");
             }
 
-            var values = new RedisValue[2 + (useApproximateMaxLength ? 1 : 0)];
+            if (limit is null && !useApproximateMaxLength && mode == StreamTrimMode.KeepReferences)
+            {
+                // avoid array alloc in simple case
+                return Message.Create(Database, flags, RedisCommand.XTRIM, key, maxLen ? StreamConstants.MaxLen : StreamConstants.MinId, threshold);
+            }
 
-            values[0] = StreamConstants.MaxLen;
+            var values = new RedisValue[2 + (useApproximateMaxLength ? 1 : 0) + (limit.HasValue ? 2 : 0) + (mode == StreamTrimMode.KeepReferences ? 0 : 1)];
+
+            var offset = 0;
+
+            values[offset++] = maxLen ? StreamConstants.MaxLen : StreamConstants.MinId;
 
             if (useApproximateMaxLength)
             {
-                values[1] = StreamConstants.ApproximateMaxLen;
-                values[2] = maxLength;
+                values[offset++] = StreamConstants.ApproximateMaxLen;
             }
-            else
+
+            values[offset++] = threshold;
+
+            if (limit.HasValue)
             {
-                values[1] = maxLength;
+                values[offset++] = RedisLiterals.LIMIT;
+                values[offset++] = limit.GetValueOrDefault();
             }
+
+            if (mode != StreamTrimMode.KeepReferences) // omit when not needed, for back-compat
+            {
+                values[offset++] = StreamConstants.GetMode(mode);
+            }
+
+            Debug.Assert(offset == values.Length);
 
             return Message.Create(
                 Database,